+++ arcade/experimental/input/__init__.py
@@ -1,0 +1,6 @@
+# ruff: noqa: F401
+#  type: ignore
+
+from .inputs import ControllerAxes, ControllerButtons, Keys, MouseAxes, MouseButtons
+from .manager import ActionState, InputManager
+from .mapping import Action, ActionMapping, Axis, AxisMapping