+++ arcade/gui/examples/widget_gallery.py
@@ -23,6 +23,7 @@ from arcade.gui import (
     UITextArea,
     UIInputText,
     UIBoxLayout,
+    UISlider,
 )
 from arcade.gui.examples.textured_slider import UITextureSlider
 from arcade.gui.widgets.layout import UIAnchorLayout
@@ -38,7 +39,7 @@ class MyView(arcade.View):
             UIGridLayout(
                 size_hint=(0.9, 0.9),
                 column_count=2,
-                row_count=6,
+                row_count=7,
                 vertical_spacing=5,
                 horizontal_spacing=5,
             )
@@ -49,15 +50,9 @@ class MyView(arcade.View):
             col_num=0,
             row_num=0,
             child=UITextureButton(
-                texture=load_texture(
-                    ":resources:gui_basic_assets/red_button_normal.png"
-                ),
-                texture_hovered=load_texture(
-                    ":resources:gui_basic_assets/red_button_hover.png"
-                ),
-                texture_pressed=load_texture(
-                    ":resources:gui_basic_assets/red_button_press.png"
-                ),
+                texture=load_texture(":resources:gui_basic_assets/red_button_normal.png"),
+                texture_hovered=load_texture(":resources:gui_basic_assets/red_button_hover.png"),
+                texture_pressed=load_texture(":resources:gui_basic_assets/red_button_press.png"),
             ),
         )
 
@@ -71,27 +66,21 @@ class MyView(arcade.View):
                     right=5,
                     bottom=5,
                     top=5,
-                    texture=load_texture(
-                        ":resources:gui_basic_assets/red_button_normal.png"
-                    ),
+                    texture=load_texture(":resources:gui_basic_assets/red_button_normal.png"),
                 ),
                 texture_hovered=NinePatchTexture(
                     left=5,
                     right=5,
                     bottom=5,
                     top=5,
-                    texture=load_texture(
-                        ":resources:gui_basic_assets/red_button_hover.png"
-                    ),
+                    texture=load_texture(":resources:gui_basic_assets/red_button_hover.png"),
                 ),
                 texture_pressed=NinePatchTexture(
                     left=5,
                     right=5,
                     bottom=5,
                     top=5,
-                    texture=load_texture(
-                        ":resources:gui_basic_assets/red_button_press.png"
-                    ),
+                    texture=load_texture(":resources:gui_basic_assets/red_button_press.png"),
                 ),
             ),
         )
@@ -112,22 +101,14 @@ class MyView(arcade.View):
         )
         toggles.add(
             UITextureToggle(
-                on_texture=load_texture(
-                    ":resources:gui_basic_assets/toggle/switch_green.png"
-                ),
-                off_texture=load_texture(
-                    ":resources:gui_basic_assets/toggle/switch_red.png"
-                ),
+                on_texture=load_texture(":resources:gui_basic_assets/toggle/switch_green.png"),
+                off_texture=load_texture(":resources:gui_basic_assets/toggle/switch_red.png"),
             )
         )
         toggles.add(
             UITextureToggle(
-                on_texture=load_texture(
-                    ":resources:gui_basic_assets/toggle/switch_green.png"
-                ),
-                off_texture=load_texture(
-                    ":resources:gui_basic_assets/toggle/switch_red.png"
-                ),
+                on_texture=load_texture(":resources:gui_basic_assets/toggle/switch_green.png"),
+                off_texture=load_texture(":resources:gui_basic_assets/toggle/switch_red.png"),
             )
         ).disabled = True
 
@@ -136,12 +117,8 @@ class MyView(arcade.View):
             col_num=0,
             row_num=4,
             child=UITextureSlider(
-                track=arcade.load_texture(
-                    ":resources:gui_basic_assets/slider_track.png"
-                ),
-                thumb=arcade.load_texture(
-                    ":resources:gui_basic_assets/slider_thumb.png"
-                ),
+                track=arcade.load_texture(":resources:gui_basic_assets/slider_track.png"),
+                thumb=arcade.load_texture(":resources:gui_basic_assets/slider_thumb.png"),
             ),
         )
 
@@ -151,19 +128,22 @@ class MyView(arcade.View):
             row_num=5,
             child=UITextureSlider(
                 track=NinePatchTexture(
-                    texture=arcade.load_texture(
-                        ":resources:gui_basic_assets/slider_track.png"
-                    ),
+                    texture=arcade.load_texture(":resources:gui_basic_assets/slider_track.png"),
                     left=30,
                     right=33,
                     bottom=18,
                     top=18,
                 ),
-                thumb=arcade.load_texture(
-                    ":resources:gui_basic_assets/slider_thumb.png"
-                ),
+                thumb=arcade.load_texture(":resources:gui_basic_assets/slider_thumb.png"),
                 height=40,
             ),
+        )
+
+        # Simple slider
+        grid.add(
+            col_num=0,
+            row_num=6,
+            child=UISlider(),
         )
 
         # Input text