+++ arcade/gui/surface.py
@@ -6,6 +6,7 @@ from typing import Tuple, Union, Optional
 import arcade
 from arcade import Texture
 from arcade.color import TRANSPARENT_BLACK
+from arcade.draw_commands import draw_lbwh_rectangle_textured
 from arcade.camera import OrthographicProjector, OrthographicProjectionData, CameraData
 from arcade.gl import Framebuffer
 from arcade.gui.nine_patch import NinePatchTexture
@@ -55,11 +56,8 @@ class Surface:
         self._cam = OrthographicProjector(
             view=CameraData(),
             projection=OrthographicProjectionData(
-                0.0, self.width,
-                0.0, self.height,
-                -100, 100,
-                (0, 0, self.width, self.height)
-            )
+                0.0, self.width, 0.0, self.height, -100, 100, (0, 0, self.width, self.height)
+            ),
         )
 
     @property
@@ -119,7 +117,7 @@ class Surface:
 
             tex.draw_sized(size=(width, height))
         else:
-            arcade.draw_lrwh_rectangle_textured(
+            draw_lbwh_rectangle_textured(
                 bottom_left_x=x,
                 bottom_left_y=y,
                 width=width,
@@ -215,4 +213,3 @@ class Surface:
         self.texture = self.ctx.texture(self.size_scaled, components=4)
         self.fbo = self.ctx.framebuffer(color_attachments=[self.texture])
         self.fbo.clear()
-